Move schema-validation to plugins folder
