+++ packages/extensions/core/src/lib/pipeline/plugins/schema-validation/index.ts
