+++ packages/extensions/core/src/lib/pipeline/plugins/schema-validation/json-schema-validator.ts
