+++ packages/extensions/core/src/lib/pipeline/plugins/schema-validation/openapi3-schema-validator.ts
