+++ packages/extensions/core/src/lib/pipeline/plugins/schema-validation/schema-validation-plugins.ts
@@ -4,9 +4,9 @@
  *--------------------------------------------------------------------------------------------*/
 import { DataHandle } from "@azure-tools/datastore";
 import { OperationAbortedException } from "@autorest/common";
-import { Channel } from "../../message";
-import { createPerFilePlugin, PipelinePlugin } from "../common";
-import { AutorestContext } from "../../configuration";
+import { Channel } from "../../../message";
+import { createPerFilePlugin, PipelinePlugin } from "../../common";
+import { AutorestContext } from "../../../configuration";
 import { SwaggerSchemaValidator } from "./swagger-schema-validator";
 import { PositionedValidationError } from "./json-schema-validator";
 import { OpenApi3SchemaValidator } from "./openapi3-schema-validator";
