+++ packages/extensions/core/src/lib/pipeline/plugins/schema-validation/swagger-schema-validator.ts
